Refactor so that numbering CLI is in numbering.py
